business3_local_seo/requirements.txt
anthropic>=0.40.0
requests>=2.31.0
fpdf2>=2.8.0
beautifulsoup4>=4.12.0
